# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Label Computation System API
  version: 1.0.0
  description: |
    High-performance label computation system for quantitative trading that processes support/resistance level events 
    and computes forward-looking labels for pattern mining and ML models.
    
    Key features:
    - Dual-mode processing: Batch backfill (minutes) and real-time incremental (<1 second)
    - Multi-timeframe alignment using lower granularity data for path-dependent calculations
    - Support for 29 FX pairs and indices with billions of labels
    
    Critical SLAs:
    - Incremental compute: <100ms p99 latency
    - Batch throughput: 1M+ candles/minute
    - Cache hit rate: >95% for recent 24 hours
  contact:
    name: Label Computation System Team
    email: support@labelcompute.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.labelcompute.com/v1
    description: Production server
  - url: https://staging-api.labelcompute.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /batch/backfill:
    post:
      tags: [Batch]
      summary: Start batch backfill operation
      description: |
        Initiate batch computation of labels for a date range.
        Target throughput: 1M+ candles/minute.
        
        Returns immediately with job ID for status tracking.
      operationId: startBatchBackfill
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchBackfillRequest'
            example:
              instrument_id: "EURUSD"
              granularity: "H4"
              start_date: "2024-01-01T00:00:00Z"
              end_date: "2024-01-31T23:59:59Z"
              label_types:
                - "enhanced_triple_barrier"
                - "vol_scaled_return"
                - "return_quantile"
              options:
                chunk_size: 10000
                parallel_workers: 8
                force_recompute: false
      responses:
        '202':
          description: Backfill job started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobResponse'
              example:
                job_id: "bf_20240110_eurusd_h4_abc123"
                status: "started"
                estimated_duration_minutes: 45
                estimated_candles: 186000
                _links:
                  self: "/batch/jobs/bf_20240110_eurusd_h4_abc123"
                  status: "/batch/jobs/bf_20240110_eurusd_h4_abc123/status"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Backfill already in progress for this range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "BACKFILL_IN_PROGRESS"
                  message: "Backfill already running for EURUSD H4"
                  details:
                    existing_job_id: "bf_20240109_eurusd_h4_def456"
  /batch/jobs/{job_id}:
    get:
      tags: [Batch]
      summary: Get batch job status
      description: Monitor the progress of a batch backfill operation
      operationId: getBatchJobStatus
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          description: Batch job identifier
          schema:
            type: string
            pattern: '^bf_[0-9]{8}_[a-z0-9]+_[a-z0-9]+_[a-z0-9]{6}$'
          example: "bf_20240110_eurusd_h4_abc123"
      responses:
        '200':
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobStatus'
              example:
                job_id: "bf_20240110_eurusd_h4_abc123"
                status: "running"
                progress:
                  completed_candles: 125000
                  total_candles: 186000
                  percentage: 67.2
                  current_date: "2024-01-21T00:00:00Z"
                performance:
                  candles_per_minute: 2750
                  avg_compute_time_ms: 22
                  cache_hit_rate: 0.94
                estimated_completion: "2024-01-10T15:30:00Z"
                created_at: "2024-01-10T14:15:00Z"
                updated_at: "2024-01-10T15:12:00Z"
        '404':
          $ref: '#/components/responses/NotFound'
  /batch/jobs:
    get:
      tags: [Batch]
      summary: List batch jobs
      description: Get paginated list of batch jobs with filtering
      operationId: listBatchJobs
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by job status
          required: false
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: instrument_id
          in: query
          description: Filter by instrument
          required: false
          schema:
            type: string
            example: "EURUSD"
        - name: granularity
          in: query
          description: Filter by granularity
          required: false
          schema:
            type: string
            enum: [M15, H1, H4, D, W]
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          description: Items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [created_at_desc, created_at_asc, updated_at_desc]
            default: created_at_desc
      responses:
        '200':
          description: Jobs list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobsList'

  # ===========================================================================
  # LABEL QUERIES
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing user claims and permissions.
        Header format: `Authorization: Bearer <token>`
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for service-to-service authentication.
        Header format: `X-API-Key: <key>`
  schemas:
...
    BatchBackfillRequest:
      type: object
      required: [instrument_id, granularity, start_date, end_date]
      properties:
        instrument_id:
          type: string
          pattern: '^[A-Z]{6}$|^[A-Z0-9]+$'
          example: "EURUSD"
        granularity:
          type: string
          enum: [M15, H1, H4, D, W]
        start_date:
          type: string
          format: date-time
          description: Start date (inclusive)
        end_date:
          type: string
          format: date-time
          description: End date (exclusive)
        label_types:
          type: array
          items:
            type: string
          description: Label types to compute (all if empty)
        options:
          type: object
          properties:
            chunk_size:
              type: integer
              minimum: 1000
              maximum: 50000
              default: 10000
              description: Candles to process per chunk
            parallel_workers:
              type: integer
              minimum: 1
              maximum: 16
              default: 8
              description: Number of parallel workers
            force_recompute:
              type: boolean
              default: false
              description: Recompute existing labels
            priority:
              type: string
              enum: [low, normal, high]
              default: normal
              description: Job priority

    # RESPONSE SCHEMAS
...
    BatchJobResponse:
      type: object
      required: [job_id, status, estimated_duration_minutes, estimated_candles]
      properties:
        job_id:
          type: string
          pattern: '^bf_[0-9]{8}_[a-z0-9]+_[a-z0-9]+_[a-z0-9]{6}$'
          description: Unique job identifier
        status:
          type: string
          enum: [started, queued]
          description: Initial job status
        estimated_duration_minutes:
          type: integer
          minimum: 1
          description: Estimated completion time in minutes
        estimated_candles:
          type: integer
          minimum: 1
          description: Total candles to process
        priority:
          type: string
          enum: [low, normal, high]
          description: Job priority
        _links:
          type: object
          description: HATEOAS links for job operations
          properties:
            self:
              type: string
              format: uri
              description: Link to job details
            status:
              type: string
              format: uri
              description: Link to job status
            cancel:
              type: string
              format: uri
              description: Link to cancel job
    BatchJobStatus:
      type: object
      required: [job_id, status, progress, created_at, updated_at]
      properties:
        job_id:
          type: string
          pattern: '^bf_[0-9]{8}_[a-z0-9]+_[a-z0-9]+_[a-z0-9]{6}$'
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled, paused]
        progress:
          type: object
          required: [completed_candles, total_candles, percentage]
          properties:
            completed_candles:
              type: integer
              minimum: 0
            total_candles:
              type: integer
              minimum: 1
            percentage:
              type: number
              format: double
              minimum: 0
              maximum: 100
            current_date:
              type: string
              format: date-time
              description: Current processing date
            chunks_completed:
              type: integer
              minimum: 0
            chunks_total:
              type: integer
              minimum: 1
        performance:
          type: object
          properties:
            candles_per_minute:
              type: number
              format: double
              minimum: 0
            avg_compute_time_ms:
              type: number
              format: double
              minimum: 0
            cache_hit_rate:
              type: number
              format: double
              minimum: 0
              maximum: 1
            error_rate:
              type: number
              format: double
              minimum: 0
              maximum: 1
        estimated_completion:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        error_message:
          type: string
          nullable: true
          description: Error message if status is failed
    BatchJobsList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BatchJobStatus'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
...
    PaginationInfo:
      type: object
      required: [page, per_page, total, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        per_page:
          type: integer
          minimum: 1
          description: Items per page
        total:
          type: integer
          minimum: 0
          description: Total items count
        total_pages:
          type: integer
          minimum: 0
          description: Total pages count
        has_next:
          type: boolean
          description: Whether next page exists
        has_prev:
          type: boolean
          description: Whether previous page exists
        next_page:
          type: integer
          nullable: true
          description: Next page number
        prev_page:
          type: integer
          nullable: true
          description: Previous page number
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid input data"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: Field name with error
                  message:
                    type: string
                    description: Field-specific error message
                  code:
                    type: string
                    description: Field-specific error code
              description: Detailed validation errors
            trace_id:
              type: string
              description: Request trace ID for debugging
              example: "req_abc123def456"
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "BAD_REQUEST"
              message: "Invalid request parameters"
              trace_id: "req_abc123def456"
    Unauthorized:
      description: Authentication required
      headers:
        WWW-Authenticate:
          description: Authentication method required
          schema:
            type: string
            example: 'Bearer realm="api", error="invalid_token"'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
              trace_id: "req_abc123def456"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "Labels not found for specified parameters"
              trace_id: "req_abc123def456"
...